Split spec Sequel models into their own files
