+++ spec/support/sequel_db.rb
@@ -1,0 +1,13 @@
+def sequel_db
+  @sequel_db = Sequel.connect(if defined?(JRUBY_VERSION)
+                                'jdbc:sqlite:sequel_data.sqlite3'
+                              else
+                                { 'adapter' => 'sqlite',
+                                  'database' => 'sequel_data.sqlite3' }
+                              end)
+end
+
+FileUtils.rm('sequel_data.sqlite3') if File.exist?('sequel_data.sqlite3')
+
+Dir["#{File.dirname(__FILE__)}/sequel_models/*.rb"].sort.each { |file| require file }
+
